+++ src/main/java/com/hwk9407/bookmanagementassignment/api/book/validator/IsbnValidator.java
@@ -1,4 +1,4 @@
-package com.hwk9407.bookmanagementassignment.domain.book;
+package com.hwk9407.bookmanagementassignment.api.book.validator;
 
 import com.hwk9407.bookmanagementassignment.exception.InvalidIsbnException;
 import org.springframework.stereotype.Component;
