+++ src/test/java/com/hwk9407/bookmanagementassignment/domain/book/IsbnValidatorTest.java
@@ -1,5 +1,6 @@
 package com.hwk9407.bookmanagementassignment.domain.book;
 
+import com.hwk9407.bookmanagementassignment.api.book.validator.IsbnValidator;
 import com.hwk9407.bookmanagementassignment.exception.InvalidIsbnException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
